move test helpers out of src
